+++ src/lib/auth-components/navigation/mobile-slider.svelte
@@ -33,7 +33,7 @@
 			<div class="flex items-center gap-[9px] mt-[35px]">
 				<div class="flex max-w-fit">
 					<img
-						src={$userState?.user_metadata.profileLink ?? sampleDisplayIcon}
+						src={$userState?.user_photo_link ?? sampleDisplayIcon}
 						alt="sample-display-icon"
 						class="w-[79px] h-[79px] rounded-full"
 					/>
@@ -50,11 +50,11 @@
 
 			<div class="mt-[25px]">
 				<p class="text-[16px] text-subwhite font-bold">
-					{$userState?.user_metadata.fullname}
+					{$userState?.user_fullname}
 				</p>
 
 				<div class="text-[14px] text-subwhite mt-[20px] rounded-[10px] max-h-[100px] overflow-auto">
-					<p>{$userState?.user_metadata.bio ?? 'You have no bio.'}</p>
+					<p>{$userState?.user_bio ?? 'You have no bio.'}</p>
 				</div>
 			</div>
