+++ src/routes/(authenticated)/+layout.server.ts
@@ -1,9 +1,15 @@
 import { redirect } from "@sveltejs/kit";
 import type { LayoutServerLoad } from "./$types";
+import type { PostgrestMaybeSingleResponse } from "@supabase/supabase-js";
+import type { UserReference } from "$lib/types";
 
-export const load: LayoutServerLoad = async ({ locals: { supabase, isLogged }, }) => {
+export const load: LayoutServerLoad = async ({ locals: { supabase, isLogged, getSession }, }) => {
     const loginCheck = isLogged();
-
+    const session = await getSession();
     if (loginCheck !== "has auth") redirect(302, "/");
+
+    return {
+        userData: await supabase.from("user_list_tb").select("*").eq("user_id", session?.user.id).limit(1).single() as PostgrestMaybeSingleResponse<UserReference>
+    }
 
 };
